Ajustando o tipo de Duração do Banco

+++ database/migrations/2022_06_23_230357_create_atrativos_table.php
@@ -20,7 +20,7 @@ class CreateAtrativosTable extends Migration
             $table->string('tipo',60);
             $table->string('observacoes',250);
             $table->integer('capacidade');
-            $table->integer('duracao');
+            $table->double('duracao');
             $table->timestamps();
         });
     }
